+++ apm-commons/apm-util/src/main/java/org/apache/skywalking/apm/util/matcher/FastPathMatcher.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.trace.ignore.matcher;
+package org.apache.skywalking.apm.util.matcher;
 
 public class FastPathMatcher implements TracePathMatcher {
     @Override