+++ apm-sniffer/optional-plugins/trace-ignore-plugin/src/test/java/org/apache/skywalking/apm/plugin/trace/ignore/TracePathMatcherTest.java
@@ -18,8 +18,8 @@
 
 package org.apache.skywalking.apm.plugin.trace.ignore;
 
-import org.apache.skywalking.apm.plugin.trace.ignore.matcher.FastPathMatcher;
-import org.apache.skywalking.apm.plugin.trace.ignore.matcher.TracePathMatcher;
+import org.apache.skywalking.apm.util.matcher.FastPathMatcher;
+import org.apache.skywalking.apm.util.matcher.TracePathMatcher;
 import org.junit.Assert;
 import org.junit.Test;